requirements.txt
requests>=2.25
packaging>=20.0
beautifulsoup4>=4.9
python-dotenv>=0.19
python-docx>=0.8.11
PyYAML>=5.4